+++ src/gui/toolselectorbutton.h
@@ -7,11 +7,23 @@
 #include "src/config.h"
 #include "src/gui/toolbutton.h"
 
+/**
+ * @brief Button showing a tool in ToolSelectorWidget
+ *
+ * This button shows a tool. Clicking the button will either set the
+ * tool for the device(s) defined in tool, or set the tool for the
+ * device used to click the button.
+ *
+ * @inherits ToolButton
+ * @see ToolSelectorWidget
+ */
 class ToolSelectorButton : public ToolButton
 {
     Q_OBJECT
 public:
+    /// Almost trivial constructor.
     ToolSelectorButton(Tool* tool, QWidget *parent = nullptr) noexcept;
+    /// Trivial destructor.
     virtual ~ToolSelectorButton() {}
 protected:
     /// Emit sendTool based on input event with adjusted device.