+++ src/gui/toolwidgetbutton.h
@@ -7,17 +7,34 @@
 #include "src/config.h"
 #include "src/gui/toolbutton.h"
 
+/**
+ * @brief Button showing a tool in ToolWidget.
+ *
+ * Shows a tool which is currently connected to a fixed input device.
+ * Clicking the button opens a dialog to change the tool.
+ *
+ * @inherits ToolButton
+ * @see ToolWidget
+ */
 class ToolWidgetButton : public ToolButton
 {
     Q_OBJECT
+    /// Input device, for which the tool is shown.
     const int device;
 
 public:
+    /// Almost trivial constructor.
     ToolWidgetButton(Tool *tool, const int device, QWidget *parent = nullptr) noexcept;
+    /// Trivial desctructor.
     virtual ~ToolWidgetButton() {}
+
 protected slots:
+    /// Select a new tool using ToolDialog.
     virtual void selectTool();
+
 public slots:
+    /// Receive a new tool from master. Only changes the tool if the
+    /// new tool is connected to this->device.
     void receiveNewTool(const Tool *newtool);
 };